Huge commit, very close to Version 1.
In this commit there were several changes made towards creating a
functioning program with hopefully minimal bugs. There is now a master
loop utilizing while )true( with a way to exit the console. A nest for
loop was used to ask the user questions in a randomized order each time
they indicate to start their test. Placeholders for further ascii
telemetry waves in use for testing randomizing questions.