Passable Version of Telemetry Testing
This commit include the first "complete" dictionary of questions for a
telemetry test. Further, there is now functionality to inform the user of
the correct answer if a question is answered incorrectly. Small changes
made to clear screen for cleaner viewing of questions. Also added a nice
title screen with thanks to Michael Winter for inspiring it. The window
size no longer extends but allows for a scroll bar at bottom of window
to view questions in their entirety. Added small changes to possible
user inputs. The test now informs the user that they have answered all
questions before returning to title menu.