Changed Up User Save Data
Now instead of using a very disorganized mix of if else statements, they are now organized under while statements with corresponding counters to organize them. The telemetry program is now accessible and uses the users name to welcome them. Can now also write to a file that records a user's score as well as the date and time they took it and how long it took them to take.