Began Work On Viewing Scores For Particular User
There is now an option to press 4 to view the date, time, and score of a previously taken test. At this time, there is still some room for error if the user does not enter in a number matching the test. As well, not sure if the code can handle if the user's scores are null i.e. they have yet to take a test.