Class CSVWriterNew and CSVSaveReader
Began work on creating classes that read and write to .csv files. CSVWriterNew can create a .csv files and append but not overwrite if one already exists. At this time, CSVSaveReader can read a .csv file and return a list of selected strings in an array of lines read from the file.