Methods Added To Class User
Added methods to the User class that will read data from a .csv file and display previous users. If the file is null or only whitespaces the class can now handle that error. Properties for an instantiated user are now saved to memory (I think that's the terminology, I just know that I can use Console.WriteLine() to display user properties)