Added More Classes And "Finished" Others
Added UserScore class to read a file specific to a user and write to it as well. Added ViewScores class to serve as the object with the properties of what a test would have that is relevent to query. Added ScoreLinq class to be used as a means for the user to order their scores by some criteria that I have yet to determine. I have "finished" the User and UserSaveData class but I'm sure there's more I could add to them, however I believe them to be in a workable state.