Added a lot of comments.
Took out .gitignore. I don't understand it and it's bad practice to try and use it and hope for the best without proper understanding. Added comments for clarity as well as where I believe a feature is being implemented. Further, included in this commit is a bin with information to test some of the features.